refactor(ui): move selected card positioning logic into a hook
